#127 making most of Task00782CleanDataInconsistencies optional
Former-commit-id: 5270511cff199756efc4099dcfb9c539a5810268

+++ src/com/dotmarketing/startup/runonce/Task00782CleanDataInconsistencies.java
@@ -7,6 +7,7 @@ import com.dotmarketing.db.DbConnectionFactory;
 import com.dotmarketing.exception.DotDataException;
 import com.dotmarketing.exception.DotRuntimeException;
 import com.dotmarketing.startup.StartupTask;
+import com.dotmarketing.util.Config;
 import com.dotmarketing.util.Logger;
 
 /**
@@ -29,35 +30,33 @@ public class Task00782CleanDataInconsistencies implements StartupTask {
             Logger.warn(this.getClass(), e.getMessage(),e);
         }
         
-        
-        final String[] smts=new String[] {
-        "delete from inode where identifier not in (select inode from identifier)",
-        "delete from file_asset where inode not in (select inode from inode)",
-        "delete from contentlet where inode not in (select inode from inode)",
-        "delete from containers where inode not in (select inode from inode)",
-        "delete from htmlpage where inode not in (select inode from inode)",
-        "delete from links where inode not in (select inode from inode)",
-        "delete from template where inode not in (select inode from inode)",
-        "delete from inode where type='folder' and not exists (select * from folder where folder.inode=inode.inode)",
-        "delete from inode where type='contentlet' and not exists (select * from contentlet where contentlet.inode=inode.inode)",
-        "delete from inode where type='file_asset' and not exists (select * from file_asset where file_asset.inode=inode.inode)",
-        "delete from inode where type='containers' and not exists (select * from containers where containers.inode=inode.inode)",
-        "delete from inode where type='template' and not exists (select * from template where template.inode=inode.inode)",
-        "delete from inode where type='htmlpage' and not exists (select * from htmlpage where htmlpage.inode=inode.inode)",
-        "delete from inode where type='containers' and not exists (select * from containers where containers.inode=inode.inode)",
-        "delete from inode where type='links' and not exists (select * from links where links.inode=inode.inode)",
-        "delete from identifier where not exists (select * from inode where inode.identifier=identifier.inode)",
-        "delete from dist_reindex_journal"
-        };
-        
-        for(String smt : smts) {
-            try {
-                dc.executeStatement(smt);
-            } catch (SQLException e) {
-                Logger.warn(this.getClass(), e.getMessage());
+        if(Config.getBooleanProperty("upgrade-cleanup-bad-data",true)) {
+            final String[] smts=new String[] {
+            "delete from inode where type='folder' and not exists (select * from folder where folder.inode=inode.inode)",
+            "delete from inode where type='contentlet' and not exists (select * from contentlet where contentlet.inode=inode.inode)",
+            "delete from inode where type='file_asset' and not exists (select * from file_asset where file_asset.inode=inode.inode)",
+            "delete from inode where type='containers' and not exists (select * from containers where containers.inode=inode.inode)",
+            "delete from inode where type='template' and not exists (select * from template where template.inode=inode.inode)",
+            "delete from inode where type='htmlpage' and not exists (select * from htmlpage where htmlpage.inode=inode.inode)",
+            "delete from inode where type='containers' and not exists (select * from containers where containers.inode=inode.inode)",
+            "delete from inode where type='links' and not exists (select * from links where links.inode=inode.inode)",
+            "delete from identifier where not exists (select * from inode where inode.identifier=identifier.inode)"
+            };
+            
+            for(String smt : smts) {
+                try {
+                    dc.executeStatement(smt);
+                } catch (SQLException e) {
+                    Logger.warn(this.getClass(), e.getMessage());
+                }
             }
         }
         
+        try {
+            dc.executeStatement("delete from dist_reindex_journal");
+        } catch(SQLException e) {
+            Logger.warn(this, "can't clean dist_reindex_journal");
+        }
     }
     
 }
